long -> double

+++ src/net/java_school/bank/Bank.java
@@ -13,11 +13,11 @@ public interface Bank extends Serializable {
 
 	public List<Account> getAccounts();
 
-	public void deposit(String accountNo, long amount);
+	public void deposit(String accountNo, double amount);
 
-	public void withdraw(String accountNo, long amount);
+	public void withdraw(String accountNo, double amount);
 
-	public void transfer(String from, String to, long amount);
+	public void transfer(String from, String to, double amount);
 
 	public List<Transaction> getTransactions(String accountNo);
